@@ -22,15 +22,17 @@
     },
     "autoload": {
         "psr-4": {
-             "Comodojo\\Zip\\": "src"
+             "Comodojo\\Zip\\": "src/Comodojo/Zip"
          }
     },
     "require": {
-        "php": ">=5.6.0",
-        "comodojo/exceptions": "@stable"
+        "php": ">=7.1",
+        "comodojo/foundation": "dev-master",
+        "ext-zip": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.0|^5.0",
+        "phpunit/phpunit": "^6.0",
         "scrutinizer/ocular": "^1.0"
-    }
+    },
+    "minimum-stability": "dev"
 }
